requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
django-admin>=1.3.2
djangorestframework>=3.9.2
djangorestframework-jwt>=1.11.0
